+++ src/android/notification/Manager.java
@@ -31,7 +31,6 @@ import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
@@ -121,7 +120,7 @@ public class Manager {
                 notification.getOptions().getDict(), updates);
 
         try {
-            options.putOpt("updatedAt", new Date().getTime());
+            options.putOpt("updated", true);
         } catch (JSONException ignore) {}
 
         return schedule(options, receiver);